refactor: modularize service registration via extensions
Refactor Program.cs by moving service registration logic into
extension methods under TournamentAPI.Configuration.Extensions.
Each concern (options, database, rate limiting, telemetry,
authentication, health checks, authorization, metrics, GraphQL,
and forwarded headers) now has its own extension method for
improved modularity and maintainability. No functional changes
were made.
