+++ TournamentAPI/Configuration/Extensions/ForwardedHeadersExtensions.cs
@@ -1,0 +1,19 @@
+using Microsoft.AspNetCore.HttpOverrides;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace TournamentAPI.Configuration.Extensions;
+
+internal static class ForwardedHeadersExtensions
+{
+    internal static IServiceCollection AddForwardedHeaders(this IServiceCollection services)
+    {
+        services.Configure<ForwardedHeadersOptions>(options =>
+        {
+            options.ForwardedHeaders = ForwardedHeaders.XForwardedFor | ForwardedHeaders.XForwardedProto;
+            options.KnownNetworks.Clear();
+            options.KnownProxies.Clear();
+        });
+
+        return services;
+    }
+}
